requirements.txt
absl-py==2.1.0
attrs==24.2.0
basemap==1.4.1
basemap-data==1.3.2
cachetools==5.5.0
cads-api-client==1.4.3
Cartopy==0.23.0
cdsapi==0.7.3
certifi==2024.8.30
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4
charset-normalizer==3.3.2
chex==0.1.87
click==8.1.7
cloudpickle==3.0.0
colabtools==0.0.1
contourpy==1.3.0
crc32c==2.7.1
cycler==0.12.1
dask==2024.9.1
Deprecated==1.2.15
dinosaur-dycore==1.2.0
dm-haiku==0.0.12
dm-tree==0.1.8
donfig==0.8.1.post1
eccodes==2.40.0
einops==0.8.0
etils==1.9.4
filelock==3.16.1
findlibs==0.1.0
flax==0.9.0
flexcache==0.3
flexparser==0.4
fonttools==4.54.1
fsspec==2024.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphcast @ https://github.com/deepmind/graphcast/archive/master.zip
huggingface-hub==0.25.1
humanize==4.10.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
jmp==0.0.4
joblib==1.4.2
jraph==0.0.6.dev0
kiwisolver==1.4.7
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MetPy==1.6.3
microsoft-aurora==1.4.1
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.0
multiurl==0.3.1
nest-asyncio==1.6.0
netCDF4==1.7.1.post2
networkx==3.3
numcodecs==0.15.0
numpy==1.26.4
opt_einsum==3.4.0
optax==0.2.3
orbax-checkpoint==0.6.4
packaging==23.2
pandas==2.2.3
partd==1.4.2
pillow==10.4.0
Pint==0.24.4
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
pysolar==0.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
rsa==4.9
Rtree==1.3.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sympy==1.13.3
tabulate==0.9.0
tensorstore==0.1.66
threadpoolctl==3.5.0
timm==0.6.13
toolz==1.0.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
tree-math==0.2.1
trimesh==4.4.9
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.17.2
xarray==2024.9.0
xarray-tensorstore==0.1.5
zarr==3.0.1
zipp==3.20.2
